Courses frontend is working completely

+++ server/handlers/department/departmentHandler.go
@@ -34,16 +34,16 @@ func CreateCourseHandler(c echo.Context, repo *repository.Repository, authClient
 	ctx := context.Background()
 	var course models.Course
 
-	// authHeader := c.Request().Header.Get("Authorization")
-	// if authHeader == "" {
-	// 	return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
-	// }
-	// tokenString := authHeader[len("Bearer "):]
-	// isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
+	authHeader := c.Request().Header.Get("Authorization")
+	if authHeader == "" {
+		return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
+	}
+	tokenString := authHeader[len("Bearer "):]
+	isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
 
-	// if !isAuth {
-	// 	return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
-	// }
+	if !isAuth {
+		return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
+	}
 
 	// Bind json data to course object
 	if err := c.Bind(&course); err != nil {
@@ -63,22 +63,21 @@ func CreateCourseHandler(c echo.Context, repo *repository.Repository, authClient
 func DeleteCourse(c echo.Context, repo *repository.Repository, authClient *auth.Client) error {
 	ctx := context.Background()
 
-	// authHeader := c.Request().Header.Get("Authorization")
-	// if authHeader == "" {
-	// 	return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
-	// }
-	// tokenString := authHeader[len("Bearer "):]
-	// isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
+	authHeader := c.Request().Header.Get("Authorization")
+	if authHeader == "" {
+		return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
+	}
+	tokenString := authHeader[len("Bearer "):]
+	isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
 
-	// if !isAuth {
-	// 	return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
-	// }
+	if !isAuth {
+		return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
+	}
 
 	courseID := c.Param("id")
 	if courseID == "" {
 		return c.JSON(http.StatusBadRequest, map[string]string{"error": "Course ID is required"})
 	}
-
 	err := repo.DeleteCourseByID(ctx, courseID)
 	if err != nil {
 		return c.JSON(http.StatusInternalServerError, map[string]string{"error": "Failed to delete course"})
@@ -122,16 +121,16 @@ func RemoveTAFromCourse(c echo.Context, repo *repository.Repository, authClient 
 func GetNewForms(c echo.Context, repo *repository.Repository, authClient *auth.Client) error {
 	ctx := context.Background()
 
-	// authHeader := c.Request().Header.Get("Authorization")
-	// if authHeader == "" {
-	// 	return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
-	// }
-	// tokenString := authHeader[len("Bearer "):]
-	// isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
+	authHeader := c.Request().Header.Get("Authorization")
+	if authHeader == "" {
+		return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
+	}
+	tokenString := authHeader[len("Bearer "):]
+	isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
 
-	// if !isAuth {
-	// 	return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
-	// }
+	if !isAuth {
+		return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
+	}
 
 	formsList, err := repo.GetNewForms(ctx)
 	if err != nil {
@@ -145,16 +144,16 @@ func GetNewForms(c echo.Context, repo *repository.Repository, authClient *auth.C
 func GetFormById(c echo.Context, repo *repository.Repository, authClient *auth.Client) error {
 	ctx := context.Background()
 
-	// authHeader := c.Request().Header.Get("Authorization")
-	// if authHeader == "" {
-	// 	return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
-	// }
-	// tokenString := authHeader[len("Bearer "):]
-	// isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
+	authHeader := c.Request().Header.Get("Authorization")
+	if authHeader == "" {
+		return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
+	}
+	tokenString := authHeader[len("Bearer "):]
+	isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
 
-	// if !isAuth {
-	// 	return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
-	// }
+	if !isAuth {
+		return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
+	}
 
 	formID := c.Param("id")
 	if formID == "" {
@@ -238,16 +237,16 @@ func UpdateDepartmentForm(c echo.Context, repo *repository.Repository, authClien
 	var form models.Form
 	ctx := context.Background()
 
-	// authHeader := c.Request().Header.Get("Authorization")
-	// if authHeader == "" {
-	// 	return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
-	// }
-	// tokenString := authHeader[len("Bearer "):]
-	// isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
+	authHeader := c.Request().Header.Get("Authorization")
+	if authHeader == "" {
+		return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
+	}
+	tokenString := authHeader[len("Bearer "):]
+	isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
 
-	// if !isAuth {
-	// 	return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
-	// }
+	if !isAuth {
+		return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
+	}
 
 	// Bind json data to form object
 	if err := c.Bind(&form); err != nil {
@@ -265,15 +264,15 @@ func UpdateDepartmentForm(c echo.Context, repo *repository.Repository, authClien
 func GetDepartmentFormsByA(c echo.Context, repo *repository.Repository, authClient *auth.Client) error {
 	ctx := context.Background()
 
-	// authHeader := c.Request().Header.Get("Authorization")
-	// if authHeader == "" {
-	// 	return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
-	// }
-	// tokenString := authHeader[len("Bearer "):]
-	// isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
-	// if !isAuth {
-	// 	return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
-	// }
+	authHeader := c.Request().Header.Get("Authorization")
+	if authHeader == "" {
+		return c.JSON(http.StatusUnauthorized, map[string]string{"error": "Missing Authorization header"})
+	}
+	tokenString := authHeader[len("Bearer "):]
+	isAuth, authMessage := AuthUser(ctx, tokenString, repo, authClient)
+	if !isAuth {
+		return c.JSON(http.StatusBadRequest, map[string]string{"error": authMessage})
+	}
 
 	taID := c.Param("taId")
 	if taID == "" {
